refactor: extract the API server lifecycle into its own file
Pure code motion, no behaviour change: StartWithContext, the three start*
helpers, startAwsLambda, lambdaHandler, Stop and isMonitoringPortConfigured
move from server.go into server_lifecycle.go.

Echo v5 removes Echo.Start/StartServer/Shutdown/Listener in favour of
echo.StartConfig, so this is the one genuinely semantic change in the
upcoming migration. Isolating it now keeps that diff reviewable instead of
burying it in server.go alongside route and middleware wiring.

Preparation for #5294.
